% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{goc}
\alias{goc}
\title{Leading NA}
\usage{
goc(n_in, l_h, n_h, n_out, Ws, Fin, flen, t_eps, t_steps)
}
\arguments{
\item{x}{An integer vector}
}
\description{
This function returns a logical vector identifying if
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}
